Added usage for slash command, status command
It occurred to me that I may want to be able to check my status of guild
repairs without having to toggle on and off, so added a status command.

I also thought about the way cli commands work in Linux and I thought it
might be a good idea to have a usage command for those not familiar with
command line environments.